requirements_dev.txt
-r requirements.txt

codecov==2.1.10
coloredlogs==14.0
coverage==5.3
django-naomi==0.8
factory_boy==2.8.1
faker==4.18.0
flake8-blind-except==0.1.1
flake8-commas==2.0.0
flake8-docstrings==1.5.0
flake8-isort==4.0.0
flake8-pep3101==1.3.0
flake8-quotes==3.2.0
flake8-string-format==0.3.0
flake8==3.8.4
ipython==7.16.3
isort==5.6.4
mock==4.0.2
parameterized==0.8.1
pydocstyle==5.1.1
pytest-cov==2.10.1
pytest-django==4.1.0
pytest==6.1.2
python-coveralls==2.9.3
sphinx==3.3.0
autopep8==1.5.4
